fix routing welcome page as homepage
