+++ pages/api/auth/[...nextauth].js
@@ -42,7 +42,7 @@ export const authOptions = {
   },
   callbacks: {
     async redirect({ url, baseUrl }) {
-      return `${baseUrl}/welcome`;
+      return `${baseUrl}/`;
     },
   },
 };
